refactor(types): move ProjectionType to shared types directory
