+++ apps/frontend/src/types/projection.ts
@@ -1,0 +1,1 @@
+export type ProjectionType = 'mercator' | 'globe';
